Put getter functions into separated file
